Address new(ish) rate limiting rules
* Closes #33
* Makes SlackClient more 429 tolerant
* Decreases overall api request speed :(

What is going on
-----

Slack has [threatened](https://api.slack.com/changelog/2018-03-great-rate-limits), [implemented](https://api.slack.com/docs/rate-limits), and finally expanded its rate limiting to the emoji api endpoints. This means some of our heavier requests (see: uploading a shit ton of emoji as fast as possible) are catching 429's and generally having a bad time.

What we should do about it
-----

There are already some features in place for this, namely the inclusion of the superagent-throttle plugging. It works, but needs to be tuned to the endpoints we use to be respectful to the slack api. What tuning does it need? Well... I don't know. Because emojme is running on undocumented endpoints, the endpoints don't have an obvious "Tier" associated with them (refer to https://api.slack.com/docs/rate-limits for what I mean by tiers).

So what tier are we?
-----

Hard to say, as we can only (as of 02/10/2019) guess by approximating at this point, but here is my basis for thinking the emoji endpoints are **Tier 3**:

* Testing
    * Having set the throttler to Tier 4, with 100 concurrent requests per 60000ms, we're still seeing 429's. Que lastima.
* Comparisons
  * Tier 2 (20 req / minute) methods seem to refer to much heavier methods than an emoji add, which I assume to be at the same logic level as adding a message (though those are no doubt more prioritized), file (probably more similar), or post.
    * emoji.list (:grimace:) this one is _unpaginated_ and a db call, so it's gonna be pretty restricted
    * search.messages is a _heavy_ db call
    * files.upload _is_ a Tier 2 method, and may be similar to emoji, however I think emoji would be "easier" on slack due to size requirements.